creation des tiles en recuperant coordonnées des pixels + la texture qui correspond + la rotation qui faudra faire lors du rendu. Il ne manque plus que de rendre les textures avec OpenGL

+++ include/Node.hpp
@@ -28,5 +28,5 @@ public:
     vector<int> getNeighbors();
     void setStatus(NodeStatus status);
     void addNeighbor(int neighbor);
-    double distanceBetweenNodes(Node target);
+    int distanceBetweenNodes(Node target);
 };
